RISC-V: Fix readelf's R_RISCV_{ADD,SUB}{8,16,32,64} handling
I'm not entirely sure this is correct, but it does pass the test cases
(as well as a handful of small modifications I made to them).  The
implementation is patterned off the existing ones, with the additional
wrinkle that we have more relocations than the other ports do.

It feels a bit odd to handle the relocations this way, as I'd expect
this to either be handled by BFD or at least to look up the relocations
via address (as opposed to just assuming tehy're in a particular order),
but I can't figure out a better way to do this.

binutils/ChangeLog

2017-10-16  Palmer Dabbelt  <palmer@dabbelt.com>

        * readelf.c (target_specific_reloc_handling) <EM_RISCV>: Handle
        R_RISCV_{ADD,SUB}{8,16,32,64} relocations.

gas/ChangeLog

2017-10-16  Palmer Dabbelt  <palmer@dabbelt.com>

        * testsuite/gas/lns/lns.exp: Use lns-common-1-alt instead of
        lns-common-1 for RISC-V targets.